+++ lss/powspec_periodic_box.cpp
@@ -12,7 +12,7 @@
 using namespace std;
 
 int Ngrid = 512;
-int Ngrid3 = Ngrid*Ngrid*Ngrid;
+long Ngrid3 = Ngrid*Ngrid*Ngrid;
 int nbin = 50; //for P(k)
 double Lbox = 1500;
 
@@ -28,13 +28,13 @@ double sinc(double x)
 }
 
 template <class T>
-T* initialize_vector(int dim)
+T* initialize_vector(long dim)
 {
 
   T *x;
   x = new T[dim];
 
-  for(int i=0;i<dim;i++) x[i] = 0.0;
+  for(long i=0;i<dim;i++) x[i] = 0.0;
 
   return x;
 }
@@ -70,31 +70,31 @@ double* initialize_fft_grid_vectors2()
 }
 
 template <class V>
-V*** initialize_3D_array(int dim1, int dim2, int dim3)
+V*** initialize_3D_array(long dim1, long dim2, long dim3)
 {
 
   V ***x;
 
   x = new V**[dim1];
 
-  for(int i=0;i<dim1;i++)
+  for(long i=0;i<dim1;i++)
   {
     x[i] = new V*[dim2];
-    for(int j=0;j<dim2;j++)
+    for(long j=0;j<dim2;j++)
     {
       x[i][j] = new V[dim3];
-      for(int k=0;k<dim3;k++) x[i][j][k] = 0.0;
+      for(long k=0;k<dim3;k++) x[i][j][k] = 0.0;
     }
   }
 
   return x;
 }
 
-int kill_3D_array(double ***x, int dim1, int dim2)
+int kill_3D_array(double ***x, long dim1, long dim2)
 {
-  for(int i=0;i<dim1;i++)
+  for(long i=0;i<dim1;i++)
   {
-    for(int j=0;j<dim2;j++)
+    for(long j=0;j<dim2;j++)
     {
       delete[] x[i][j];
     }
@@ -198,7 +198,7 @@ int powspec(double ***delta, double *km, double *pk)
 
 //Do fft to get P(k)...
   double *delta_in;
-  int r2csize = Ngrid*Ngrid*(Ngrid+2);
+  long r2csize = Ngrid*Ngrid*(Ngrid+2);
   delta_in = initialize_vector<double>(r2csize);
   int count=0;
   for(int i=0;i<Ngrid;i++)
@@ -244,8 +244,8 @@ int powspec(double ***delta, double *km, double *pk)
         //for some reason I need to divide by Ngrid**3...this appears to be due
         //to a difference between how the FFT is defined in fftw versus IDL...
         //not sure which one is correct????
-        int redex = count;
-        int imdex = count+1;
+        long redex = count;
+        long imdex = count+1;
         sinckz = sinc(kgrid[k]*lcico2);
         sinckz *= sinckz;
         wcic = sinckx*sincky*sinckz;
@@ -293,7 +293,7 @@ int crossspec(double ***d1, double ***d2, double *km, double *pk)
 
 //Do fft to get P(k)...
   double *delta_1, *delta_2;
-  int r2csize = Ngrid*Ngrid*(Ngrid+2);
+  long r2csize = Ngrid*Ngrid*(Ngrid+2);
   delta_1 = initialize_vector<double>(r2csize);
   delta_2 = initialize_vector<double>(r2csize);
   int count=0;
@@ -346,8 +346,8 @@ int crossspec(double ***d1, double ***d2, double *km, double *pk)
         //for some reason I need to divide by Ngrid**3...this appears to be due
         //to a difference between how the FFT is defined in fftw versus IDL...
         //not sure which one is correct????
-        int redex = count;
-        int imdex = count+1;
+        long redex = count;
+        long imdex = count+1;
         sinckz = sinc(kgrid[k]*lcico2);
         sinckz *= sinckz;
         wcic = sinckx*sincky*sinckz;
@@ -376,23 +376,23 @@ int crossspec(double ***d1, double ***d2, double *km, double *pk)
   return 0;
 }
 
-int cic_mass(_gal *p, int N, double ***rho)
+int cic_mass(_gal *p, long N, double ***rho)
 {
   double lside = Ngrid/Lbox;
 
-  for(int i=0;i<N;i++)
+  for(long i=0;i<N;i++)
   {
 //Because the box is periodic, the Ngrid+1th grid point is the 0th grid point!
 //That's why Ngrid == Ncell!
     double dd1 = p[i].x*lside;
     double dd2 = p[i].y*lside;
     double dd3 = p[i].z*lside;
-    int i1 = floor(dd1);
-    int i2 = floor(dd2);
-    int i3 = floor(dd3);
-    int j1 = i1 + 1;
-    int j2 = i2 + 1;
-    int j3 = i3 + 1;
+    long i1 = floor(dd1);
+    long i2 = floor(dd2);
+    long i3 = floor(dd3);
+    long j1 = i1 + 1;
+    long j2 = i2 + 1;
+    long j3 = i3 + 1;
     dd1 = dd1 - i1;
     dd2 = dd2 - i2;
     dd3 = dd3 - i3;
@@ -440,7 +440,7 @@ int cic_mass(_gal *p, int N, double ***rho)
   return 0;
 }
 
-int calc_delta(double ***delta, _gal *p, int N)
+int calc_delta(double ***delta, _gal *p, long N)
 {
   double ***rho;
   rho = initialize_3D_array<double>(Ngrid,Ngrid,Ngrid);
@@ -468,9 +468,9 @@ int calc_delta(double ***delta, _gal *p, int N)
   return 0;
 }
 
-int get_num_lines(string infile)
+long get_num_lines(string infile)
 {
-  int nline=0;
+  long nline=0;
   string line;
   ifstream ff;
 
@@ -490,7 +490,7 @@ int get_num_lines(string infile)
   return nline-1;
 }
 
-_gal* read_file(string fname, int *N)
+_gal* read_file(string fname, long *N)
 {
   //Read in galaxies
   ifstream infile;
@@ -503,7 +503,7 @@ _gal* read_file(string fname, int *N)
   g = new _gal[*N];
   infile.open( fname.c_str(), ios::in );
 
-  for(int i=0;i<*N;i++)
+  for(long i=0;i<*N;i++)
   {
     string line;
 
@@ -533,7 +533,7 @@ int main(int argc, char* argv[])
   }
 
   _gal *gal;
-  int Ngal;
+  long Ngal;
   gal = read_file(infile+".txt",&Ngal);
   cout << "Galaxies:" << endl;
   cout << gal[0].x << ' ' << gal[0].y << ' ' << gal[0].z << endl;
